[Refactor] 좋아요 기능 로직 변경 - 반정규화 : Post에 likeCount 기능 추가 - 긍정적인 반응이면 likeCount를 증가시킨다. - 유저가 이미 존재하는 같은 타입의 리액션을 보내면 삭제한다. - visitCount Long -> int

+++ src/main/java/store/ppingpong/board/post/domain/PostWithImages.java
@@ -19,13 +19,13 @@ public class PostWithImages {
     private final Long postTypeId;
     private final Long userId;
     private final String forumId;
-    private final Long visitCount;
+    private final int visitCount;
     private final int likeCount;
     private final int dislikeCount;
     private final LocalDateTime createdAt;
 
     @Builder
-    private PostWithImages(Long postId, String title, String content, List<Image> images, Long postTypeId, Long userId, String forumId, Long visitCount, int likeCount, int dislikeCount, LocalDateTime createdAt) {
+    private PostWithImages(Long postId, String title, String content, List<Image> images, Long postTypeId, Long userId, String forumId, int visitCount, int likeCount, int dislikeCount, LocalDateTime createdAt) {
         this.postId = postId;
         this.title = title;
         this.content = content;
